Added scripts to use citekeys instead of titles for item lookup.
The original script uses titles for item lookup, which has issues if there are multiple entries in the bibliography that share a title. This can happen for a few reasons, including:

1. Duplicate entries in Papers
2. Papers that happen to have the same title, which does happen, especially in the case of...
3. Papers that are first published at a conference, and then later are expanded into a full journal publication, or included in a special issue of a journal that publishes selected papers from a conference.

Fortunately, the export/import method prescribed in `Papers3_to_Zotero.py` will result in Zotero preserving citekeys, and it turns out that Papers 3's AppleScript support allows us to access those citekeys, as well. Using these allows for a much more reliable duplication of the Papers 3 collection structure and membership.